ARCHER-ROVER/requirements.txt
flask
requests
opencv-python
numpy
pyserial
